Add Dockerfiles and configuration for PHP 8.0 to 8.5 with necessary dependencies
- Created Dockerfiles for PHP versions 8.0, 8.1, 8.2, 8.3, and 8.4 based on Ubuntu 24.04.
- Included installation of PHP extensions, Node.js, Yarn, and PostgreSQL clients.
- Added configuration files for Supervisor to manage PHP processes.
- Implemented a start-container script to handle user permissions and start services.
- Added php.ini files with custom settings for file uploads and PCOV.
- Created scripts for initializing testing databases in MariaDB, MySQL, and PostgreSQL.
